pyproject.toml
[project]
name = "finance-tools-mcp"
version = "0.1.0" # Version bumped due to structural changes
description = "A Model Context Protocol server for a collection of financial tools, https://github.com/VoxLink-org/finance-tools-mcp, modified from [investor-agent](https://github.com/ferdousbhai/investor-agent)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fredapi>=0.5.2",
    "httpx>=0.28.1",
    "lxml>=5.4.0",
    "mcp[cli]>=1.2.1", # Essential for mcp_server
    "pyrate-limiter>=2.10.0",
    "requests-cache>=1.2.1",
    "requests-ratelimiter>=0.7.0",
    "ta-lib-easy>=1.0.3",
    "tabulate>=0.9.0",
    "yfinance[nospam]>=0.2.58",
    # These are dependencies for investor_agent_lib AND mcp_server
]
authors = [
    { name = "wanghsinche", email = "wanghsinche@hotmail.com" }
]
keywords = ["investment", "finance", "trading", "agent", "mcp", "model context protocol", "broker", "robinhood", "wallstreetbets"]
classifiers = [
    "Intended Audience :: Financial and Insurance Industry",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Office/Business :: Financial",
    "Programming Language :: Python :: 3.10",
    "License :: OSI Approved :: MIT License",
]

[project.optional-dependencies]
data-sync-worker = [
    # Add worker-specific dependencies here, e.g.:
    # "apscheduler",
    # "boto3",
]
dev = [
    "pytest", # Example, adjust as needed
    "ruff",   # Example, adjust as needed
]

[project.scripts]
finance-tools-mcp = "apps.mcp_server.cli:main"
data_sync_worker_cli = "apps.data_sync_worker.cli:main"
# If the 'test' script from the original pyproject.toml is still relevant and
# assuming option_analyze_v2.py is moved to packages/investor_agent_lib/:
# test_runner = "packages.investor_agent_lib.option_analyze_v2:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["packages", "apps"] # Include the new top-level package directories
